flask_app/requirements.txt
flask 
pandas 
bokeh 
requests 
multiprocess 
notebook 
numpy 
scipy 
OpenVisusNoGui 
GitPython
